Нурканат_Алиар_API_для_Заметок_1/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
sqlmodel==0.0.14
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.2 